Added authorizatio and Increased Robustness

+++ backend/controllers/goalControllers.js
@@ -1,12 +1,14 @@
 import asyncHandler from 'express-async-handler'
 
 import Goal from '../models/goalsModel.js';
+import User from '../models/usersModel.js';
+
 //@desc Get goals
 //@route GET /api/goals
 //@access Private
 const getGoals = asyncHandler (
     async (req, res) => {
-        const goals = await Goal.find()
+        const goals = await Goal.find({ user: req.user.id})
         res.status(200).json(goals);
     }
 ) 
@@ -22,7 +24,8 @@ const setGoal = asyncHandler (
         }
 
         const goal = await Goal.create({
-            text: req.body.text
+            text: req.body.text,
+            user: req.user.id,
         })
         res.status(200).json(goal);
     }
@@ -35,10 +38,24 @@ const setGoal = asyncHandler (
 const updateGoal = asyncHandler(
     async (req, res) => {
         const goal = await Goal.findById(req.params.id)
-
+        
         if(!goal){
             res.status(400)
             throw new Error('Goal not found');
+        }
+
+        const user = await User.findById(req.user.id)
+
+        //check for user
+        if(!user) {
+            res.status(401);
+            throw new Error('User not found');
+        }
+        
+        //Make sure the logged in user matches the goals user
+        if(goal.user.toString() !== user.id){
+            res.status(401);
+            throw new Error('User not authorized')
         }
 
         const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
@@ -57,6 +74,22 @@ const deleteGoal = asyncHandler(
             res.status(400);
             throw new Error('Goal not found');
         }
+        
+        const user = await User.findById(req.user.id)
+        
+        //check for user
+        if(!user) {
+            res.status(401);
+            throw new Error('User not found');
+        }
+        
+        //Make sure the logged in user matches the goals user
+        if(goal.user.toString() !== user.id){
+            res.status(401);
+            throw new Error('User not authorized')
+        }
+
+
 
         const deletedGoal = await Goal.findByIdAndDelete(req.params.id);
         res.status(200).json(deletedGoal);
